votecoin/requirements.txt
alembic==1.4.2
appnope==0.1.0
argh==0.26.2
asgiref==3.2.7
asn1crypto==1.3.0
atomicwrites==1.3.0
attrs==19.3.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
blinker==1.4
boto3==1.12.36
botocore==1.15.36
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
configparser==5.0.0
coverage==5.0.4
decorator==4.4.2
Django==3.0.5
docopt==0.6.2
docutils==0.16
ecdsa==0.15
entrypoints==0.3
enum34==1.1.10
flake8==3.7.9
Flask==1.1.2
Flask-API==2.0
Flask-Bcrypt==0.7.1
flask-heroku==0.1.9
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
jellyfish==0.5.6
Jinja2==2.11.1
jmespath==0.9.5
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pycountry==19.8.18
pycparser==2.20
pycryptodome==3.9.7
pyflakes==2.1.1
pyparsing==2.4.7
pytest==5.4.1
pytest-flask==1.0.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
urllib3==1.25.8
us==1.0.0
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.2.1
